requirements.txt
python-dotenv==1.0.1
Requests==2.32.3
python-telegram-bot==13.15
urllib3==1.26.15